requirements.txt
aiosqlite
bcrypt
python-dotenv
fastapi
fastapi-jwt-auth
fastapi_pagination
fastapi_mail
httpx
pydantic[email]
pyjwt
pytest
pytest_asyncio
sqlalchemy
sqlalchemy_utils
uvicorn
pytest_mock